+++ src/lib/core/cartesian-chart.ts
@@ -1,42 +1,35 @@
 import { scaleLinear } from 'd3-scale';
 import { BaseType, select, Selection } from 'd3-selection';
 import { axisBottom, axisLeft, ConfigureAxisFn, DataAxis, dataAxis } from '../axis';
+import { chart } from './chart';
 import { debug, nodeToString } from './log';
 import { ScaleAny } from './scale';
 
 export interface DataChartCartesian {
-  mainScale: ScaleAny<any, number, number>;
-  mainTitle: string;
-  mainSubtitle: string;
-  configureMainAxis: ConfigureAxisFn;
-  crossScale: ScaleAny<any, number, number>;
-  crossTitle: string;
-  crossSubtitle: string;
-  configureCrossAxis: ConfigureAxisFn;
+  mainAxis: DataAxis;
+  crossAxis: DataAxis;
+  flipped: boolean;
 }
 
 export function dataChartCartesian(data?: Partial<DataChartCartesian>): DataChartCartesian {
   return {
-    mainScale: scaleLinear().domain([0, 1]).range([0, 500]),
-    mainTitle: data?.mainTitle || '',
-    mainSubtitle: data?.mainSubtitle || '',
-    configureMainAxis: data?.configureMainAxis || (() => {}),
-    crossScale: scaleLinear().domain([0, 1]).range([0, 500]),
-    crossTitle: data?.crossTitle || '',
-    crossSubtitle: data?.crossSubtitle || '',
-    configureCrossAxis: data?.configureCrossAxis || (() => {}),
+    mainAxis: dataAxis(data?.mainAxis),
+    crossAxis: dataAxis(data?.crossAxis),
+    flipped: false,
   };
 }
 
 export function chartCartesian<
   GElement extends SVGSVGElement | SVGGElement,
-  Datum,
+  Datum extends DataChartCartesian,
   PElement extends BaseType,
   PDatum
 >(
-  selection: Selection<GElement, Datum, PElement, PDatum>
+  selection: Selection<GElement, Datum, PElement, PDatum>,
+  autoUpdateAxes: boolean
 ): Selection<GElement, Datum, PElement, PDatum> {
   return selection
+    .call((s) => chart(s))
     .classed('chart-cartesian', true)
     .each((d, i, g) => {
       const s = select<GElement, Datum>(g[i])
@@ -44,24 +37,11 @@ export function chartCartesian<
         .layout('grid-template', '1fr auto / auto 1fr')
         .layout('padding', '20px');
 
-      const drawArea = s
-        .append('svg')
+      s.append('svg')
         .classed('draw-area', true)
         .attr('overflow', 'visible')
         .layout('grid-area', '1 / 2')
         .layout('display', 'grid');
-
-      // const barSeries = drawArea
-      //   .append('g')
-      //   .layout('grid-area', '1 / 1')
-      //   .datum((d) => dataSeriesBar(d))
-      //   .call((s) => seriesBar(s));
-
-      // drawArea
-      //   .append('g')
-      //   .layout('grid-area', '1 / 1')
-      //   .datum((d) => dataSeriesLabelBar(dataLabelsBarCreation({ barContainer: barSeries })))
-      //   .call((s) => seriesLabel(s));
 
       s.append('g')
         .datum((d) => dataAxis())
@@ -73,14 +53,21 @@ export function chartCartesian<
         .call((s) => axisBottom(s))
         .layout('grid-area', '2 / 2');
     })
-    .on('datachange.chartcartesian', function (e, chartData) {
-      debug(`data change on ${nodeToString(this)}`);
-      chartCartesianDataChange(select<GElement, Datum>(this));
-    })
-    .call((s) => chartCartesianDataChange(s));
+    .call(
+      (s) =>
+        autoUpdateAxes &&
+        s
+          .on('datachange.debuglog', function () {
+            debug(`data change on ${nodeToString(this)}`);
+          })
+          .on('datachange.updateaxes', function (e, chartData) {
+            chartCartesianUpdateAxes(select<GElement, Datum>(this));
+          })
+    )
+    .call((s) => chartCartesianUpdateAxes(s));
 }
 
-export function chartCartesianDataChange<
+export function chartCartesianUpdateAxes<
   GElement extends SVGSVGElement | SVGGElement,
   Datum extends DataChartCartesian,
   PElement extends BaseType,
@@ -91,24 +78,13 @@ export function chartCartesianDataChange<
   return selection.each(function (chartData, i, g) {
     const s = select<GElement, Datum>(g[i]);
 
-    // s.selectAll<SVGElement, DataSeriesBar>('.series-bar').datum((d) =>
-    //   Object.assign(d, { creation: chartData })
-    // );
-
     const axisConfig = (selection: Selection<Element, DataAxis>, main: boolean) =>
       selection
-        .datum((d) =>
-          Object.assign(d, {
-            scale: main ? chartData.mainScale : chartData.crossScale,
-            title: main ? chartData.mainTitle : chartData.crossTitle,
-            subtitle: main ? chartData.mainSubtitle : chartData.crossSubtitle,
-            configureAxis: main ? chartData.configureMainAxis : chartData.configureCrossAxis,
-          })
-        )
+        .datum((d) => Object.assign(d, main ? chartData.mainAxis : chartData.crossAxis))
         .classed('axis-main', main)
         .classed('axis-cross', !main);
 
-    if (chartData.orientation === Orientation.Horizontal) {
+    if (!chartData.flipped) {
       s.selectAll<SVGGElement, DataAxis>('.axis-left').call((s) => axisConfig(s, true));
       s.selectAll<SVGGElement, DataAxis>('.axis-bottom').call((s) => axisConfig(s, false));
     } else {